[core] 修复A3 场景register 失败问题



# message auto-generated for no-merge-commit merge:
!669 merge release/1.1_0514_sdma into release/1.1

[core] 修复A3 场景register 失败问题

Created-by: p3rry
Commit-by: p3rry
Merged-by: chenyz6
Description: # Pull Request Template

## Description

<!-- Provide a clear summary of the change, the problem it solves, and the technical approach. -->

## Related Issues

<!-- Replace with actual issue numbers. Example: Closes #123, Related to #456 -->

## Type of Change

- [ ] Bug fix
- [ ] New feature
- [ ] Performance optimization
- [ ] Documentation update
- [ ] Other (please describe):_______________

## Testing

<!-- Describe how the changes were tested. Include:
- Unit/integration test coverage
- Manual validation steps
- Links to CI reports (e.g., CIDA) if available
- Note any pending tests
-->
![image.png](https://raw.gitcode.com/user-images/assets/7672916/80e6067d-d3ba-4112-a005-1173da200eb6/image.png 'image.png')
## Checklist

- [ ] I have performed a self-review of my own code.  
- [ ] I have updated the documentation.  
- [ ] I have added tests to prove my changes are effective.

See merge request: Ascend/memfabric_hybrid!669

+++ src/hybm/csrc/mm/hybm_dev_legacy_segment.h
@@ -91,8 +91,8 @@ protected:
     uint64_t totalVirtualSize_{0UL};
     uint8_t *lvaBase_{nullptr}; // device lva base
     uint64_t allocatedSize_{0UL};
-    std::map<uint16_t, MemSliceStatus> slices_;
-    std::map<uint16_t, HbmExportInfo> exportMap_;
+    std::map<uint32_t, MemSliceStatus> slices_;
+    std::map<uint32_t, HbmExportInfo> exportMap_;
     std::set<uint64_t> mappedGvaMem_; // gva sets
     std::vector<HbmExportInfo> imports_;
     std::map<uint32_t, HbmExportInfo> importMap_;
